requirements.txt
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
jsonschema_pydantic==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
mcp==1.9.1
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
gradio==5.31.0
gradio_client==1.10.1
aiofiles==24.1.0
aiosqlite==0.21.0

